fix: rollback to use sync.Once, update tests
