+++ node/taskfiles/taskfile.dev.yml
@@ -1,0 +1,1 @@
+version: "3"
