+++ node/taskfiles/taskfile.local.yml
@@ -1,0 +1,7 @@
+version: "3"
+
+tasks:
+  test-db:
+    dotenv: [".env"]
+    cmds:
+      - go test ./pkg/db/tests -v
