+++ node/taskfiles/taskfile.prod.yml
@@ -1,0 +1,1 @@
+version: "3"
